Add AutoSaveProvider
